change folder schema
