--- app/visualize/__init__.py
@@ -1,8 +1,0 @@
-"""
-visualize blueprint: visualize map and enable users to choose area to change veg code
-"""
-from flask import Blueprint
-
-visualize = Blueprint('visualize', __name__)
-
-from . import views
